# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TM Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /productions:
    get:
      operationId: listProductions
      tags: [productions]
      responses:
        "200":
          description: List of productions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Production"
    post:
      operationId: createProduction
      tags: [productions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Production"
  /productions/{id}:
    delete:
      operationId: deleteProduction
      tags: [productions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # Build Projects
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    patch:
      operationId: updateProject
      tags: [projects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # Shots
...
  /projects/{projectId}/script-breakdowns:
    get:
      operationId: listScriptBreakdowns
      tags: [script-breakdowns]
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScriptBreakdown"
    post:
      operationId: createScriptBreakdown
      tags: [script-breakdowns]
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScriptBreakdownInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScriptBreakdown"
  /script-breakdowns/{id}:
    patch:
      operationId: updateScriptBreakdown
      tags: [script-breakdowns]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScriptBreakdownUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScriptBreakdown"
    delete:
      operationId: deleteScriptBreakdown
      tags: [script-breakdowns]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # Project Assets
  /projects/{projectId}/project-assets:
    get:
      operationId: listProjectAssets
      tags: [project-assets]
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectAsset"
    post:
      operationId: createProjectAsset
      tags: [project-assets]
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectAssetInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectAsset"
  /project-assets/{id}:
    delete:
      operationId: deleteProjectAsset
      tags: [project-assets]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # Leads
...
  /projects/{projectId}/resources:
    get:
      operationId: listProjectResources
      tags: [project-resources]
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectResource"
    post:
      operationId: createProjectResource
      tags: [project-resources]
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectResourceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResource"
  /project-resources/{id}:
    patch:
      operationId: updateProjectResource
      tags: [project-resources]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectResourceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResource"
    delete:
      operationId: deleteProjectResource
      tags: [project-resources]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    Production:
      type: object
      required: [id, name, date, location, crew, status, createdAt, updatedAt]
      properties:
        id: { type: string }
        name: { type: string }
        date: { type: string }
        location: { type: string }
        crew: { type: integer }
        status: { type: string }
        userId: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    ProductionInput:
      type: object
      required: [name, date, location]
      properties:
        name: { type: string }
        date: { type: string }
        location: { type: string }
        crew: { type: integer }
        status: { type: string }
...
    Project:
      type: object
      required: [id, userId, name, type, status, createdAt, updatedAt]
      properties:
        id: { type: string }
        userId: { type: string }
        name: { type: string }
        description: { type: ["string", "null"] }
        type: { type: string }
        status: { type: string }
        client: { type: ["string", "null"] }
        startDate: { type: ["string", "null"] }
        endDate: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    ProjectInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        description: { type: string }
        type: { type: string }
        status: { type: string }
        client: { type: string }
        startDate: { type: string }
        endDate: { type: string }
    ProjectUpdate:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        type: { type: string }
        status: { type: string }
        client: { type: string }
        startDate: { type: string }
        endDate: { type: string }
...
    ScriptBreakdown:
      type: object
      required: [id, projectId, userId, elementType, name, createdAt, updatedAt]
      properties:
        id: { type: string }
        projectId: { type: string }
        userId: { type: string }
        elementType: { type: string }
        name: { type: string }
        description: { type: ["string", "null"] }
        sceneReference: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    ScriptBreakdownInput:
      type: object
      required: [name, elementType]
      properties:
        elementType: { type: string }
        name: { type: string }
        description: { type: string }
        sceneReference: { type: string }
    ScriptBreakdownUpdate:
      type: object
      properties:
        elementType: { type: string }
        name: { type: string }
        description: { type: string }
        sceneReference: { type: string }
    ProjectAsset:
      type: object
      required: [id, projectId, userId, name, fileType, category, createdAt, updatedAt]
      properties:
        id: { type: string }
        projectId: { type: string }
        userId: { type: string }
        name: { type: string }
        fileType: { type: string }
        fileUrl: { type: ["string", "null"] }
        category: { type: string }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    ProjectAssetInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        fileType: { type: string }
        fileUrl: { type: string }
        category: { type: string }
        notes: { type: string }
...
    ProjectResource:
      type: object
      required: [id, projectId, userId, name, category, quantity, status, cost, createdAt, updatedAt]
      properties:
        id: { type: string }
        projectId: { type: string }
        userId: { type: string }
        name: { type: string }
        category: { type: string }
        quantity: { type: integer }
        status: { type: string }
        assignedTo: { type: ["string", "null"] }
        supplier: { type: ["string", "null"] }
        cost: { type: string }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    ProjectResourceInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        category: { type: string }
        quantity: { type: integer }
        status: { type: string }
        assignedTo: { type: string }
        supplier: { type: string }
        cost: { type: string }
        notes: { type: string }
    ProjectResourceUpdate:
      type: object
      properties:
        name: { type: string }
        category: { type: string }
        quantity: { type: integer }
        status: { type: string }
        assignedTo: { type: string }
        supplier: { type: string }
        cost: { type: string }
        notes: { type: string }