# lib/api-zod/src/generated/types/projectResource.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TM Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectResource {
  id: string;
  projectId: string;
  userId: string;
  name: string;
  category: string;
  quantity: number;
  status: string;
  /** @nullable */
  assignedTo?: string | null;
  /** @nullable */
  supplier?: string | null;
  cost: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}