# lib/api-zod/src/generated/types/projectResourceUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TM Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectResourceUpdate {
  name?: string;
  category?: string;
  quantity?: number;
  status?: string;
  assignedTo?: string;
  supplier?: string;
  cost?: string;
  notes?: string;
}